Broke monolithic file into smaller sources
